Moving custom ugen into AudioKitPlaygrounds, and out of sources
