+++ AudioKit/Common/Internals/Sporth Custom Ugens/AKCustomUgen.swift
@@ -6,8 +6,6 @@
 //  Copyright © 2017 AudioKit. All rights reserved.
 //
 
-import Foundation
-
 open class AKCustomUgen: NSObject {
   open let name: String
   open let argTypes: String // string of f for float / s for string, e.g. "fsf"
